Politeness FIX
There was an error in the robot_spider which would rewrite the set of paths to exclude each time it visited a robots.txt. For some reason I did not use the set's find function in the queen's url_check for comparing a path against a path in the exclusion list, this is now corrected.